refactor: 파일, 폴더 이름 및 구조 수정
