+++ src/api/services/auth.service.ts
@@ -1,6 +1,5 @@
 import publicClient from '@/api/clients/publicClient';
-import type { LoginPayload } from '@/types/loginPayload';
-import type { UserInfo } from '@/types/userInfo';
+import type { LoginPayload, UserInfo } from '@/api/types/auth.dto';
 
 export const authService = async (payload: LoginPayload): Promise<UserInfo> => {
   const response = await publicClient.post('/api/login', payload);
