+++ src/api/types/auth.dto.ts
@@ -1,3 +1,8 @@
+export type LoginPayload = {
+  email: string;
+  password: string;
+};
+
 export type UserInfo = {
   email: string;
   name: string;
