--- src/types/categoryCardData.ts
@@ -1,5 +1,0 @@
-export type CategoryCardData = {
-  themeId: number;
-  name: string;
-  image: string;
-};
