--- src/types/loginPayload.ts
@@ -1,4 +1,0 @@
-export type LoginPayload = {
-  email: string;
-  password: string;
-};
