--- src/types/orderForm.ts
@@ -1,11 +1,0 @@
-type RecipientForm = {
-  recipientName: string;
-  phoneNumber: string;
-  amount: number;
-};
-
-export type FormValues = {
-  message: string;
-  senderName: string;
-  recipientInfo: RecipientForm[];
-};
